requirements.txt
torch==2.5.1
pillow==11.0.0
uvicorn==0.32.1
fastapi==0.115.5
torchvision==0.20.1
python-multipart==0.0.17
fastapi==0.103.0
uvicorn==0.21.0
torch==1.13.1
numpy==1.24.1
pillow==9.3.0
dnnlib==1.0.0
legacy==1.0.0
click==8.1.3
shutil
